* info/session.c (move_to_goal_column): use size_t for long.  The window->goal_column is necessarily positive, as move_to_goal_column is called after window_get_cursor_column if window->goal_column was negative.  And the LINE_MAP structure uses size_t for columns indices.

+++ info/session.c
@@ -1245,7 +1245,7 @@ void info_prev_line (WINDOW *, int count);
 static void
 move_to_goal_column (WINDOW *window)
 {
-  long goal;
+  size_t goal;
 
   goal = window->goal_column;
   if (goal >= window->line_map.used)
